refactor: extracted allocation functions from malloc to cover both scenarios

+++ ft_malloc.h
@@ -105,7 +105,7 @@ void	*realloc(void *ptr, size_t size);
 void	show_alloc_mem();
 
 // Utils
-void	*safe_mmap(int size);
+void	*safe_mmap(size_t size);
 void	*push_to_back(void *array, void *ptr);
 void	*pop(void *array, void *ptr);
 bool	contains(void *array, void *ptr);
